Create hackrooms and languages
I'm going for a simpler way of grouping people that using room maps, in that
we'll just create groups (hackrooms) based on projects and allow people to
belong to those groups.

We'll also create languages and allow groups and people to belong to those
too, so users can put themselves into languages they're interested in and we
can display those languages on a page.

We'll also display the hackrooms on a page too.