Create backround and underline paint function
This sets the backround/border colour based on the number passed to it. We'll
use this to set a colour based on the language ID of the property, so a
hackroom we can set the underline to be @hackroom.languages.first.id and it
will pass in a number. This sass function will take that number and assign the
colour from the map.