Convert interests into tags
For interests to be more userful, this converts them to tags, which a user can
have many of AS interests, and posts and hackrooms can also have tags.

This allows us to make a show page for each tag people create, which can show
the users and hackrooms related to it.