Create user follows migration
Sets up the migration that creates the User Follows join table. This will keep
track of the relation between a user and her followers, or a user and the
people they follow.

If you could only ever follow one person, we could make this a column on the
user table for the person you're following, but since you can have lots, we
need to use a separate table to keep track of them.

All this table will consist of is a column for the follwer and a column for
the followee.

We can then interact with each of the records in this table by

a) giving each user a 'has_many: :followers' relation so you can call
user.followers to return all their followers (all users in a follower column
that matches this user).

and

b) giving each user a 'has_many: :followed_users' relation to do the reverse
and show all the people they follow.