Create user follow model
This creates the model record where we define our relations. This tells Rails
what to do when you call UserFollow.follower (return the follower for this
record).

An example of using this might be: `UserFollow.find_or_create_by(follower_id:
current_user, user_id: @user)` to find the follow record for the currently
signed in user (current_user) and the @user who's profile they're viewing if
it exists, or create it if it doesn't, which would create the follow between
those users.

This could be used to create a follow link on a profile.