Update user model to have follows/followers
This adds those relations to the user model so when you call these methods on the user, Rails knows what to do.

We have to do this in two stages, first to give the user access to the join
table record, then to give the user its followers and the people she follows
VIA those join table records.

When you call user.followers, Rails will check the has_many :followers
relation, which tells it to check the UserFollows table for all rows where its
own ID matches the user_id column and return the user on the follower column
of those rows.