Add follower count to users table
This stores the total amount of followers a user has so we can quickly get the
follower count from the user, and also sort users by most followers if we need
to.

We can use 'counter_cache' to prevent Rails hitting the database every time
it's asked to fetch this count, which improves load time.