User helper methods
The list_all_sentence method returns a list of user names linked to the user's
profile, and the to_sentence method returns them in a sentence, i.e. adds
commas and an 'and' where appropriate.

The user_follower_count uses that list_all_sentence method to return the list
of followers, naming up to three and giving an amount for the rest.

The user_following? method checks if the first user passed is following the
second user passed to it. We'll use this to see if the signed in user is
currently following the person who's profile they're viewing.