Create follow/unfollow users controller actions
This creates the actions in the users controller that correspond to the
previously created routes. Those routes automatically go to the users
controller and look for these actions. When followed, Rails will carry out the
code under the action here.

In this case the 'Follow' action will find the @user in question (because of
the before_action at the top), and then it will find the UserFollow record for
that user.

As a precaution it uses find_or_create_by just in case the record already
exists. If it does we don't want to create a duplicate. However it will still
add to the follower count, so this could be improved by returning if it is
found before it adds to the count.

We should also add a unique constraint to UserFollows so you cannot have
duplicates.