Add popularity order scope to users
This adds a scope for ordering users by highest popularity first. It also adds
that scope to the users controller index action where all users are saved to
the @users instance variable.

This will order all the users on the members page by number of follows with
the person with the most followers at the top.