Create notifications migration
This creates a table in the database for storing notifications.

A notification will have a user column (indexed for performance) that is the
user who will be notified by this.

It has a notified_by column to reference the person who triggered the
notification (the user following you, the person who liked your think or
joined your hackroom).

It has a string for the action, e.g. 'followed' or 'joined'.

The notifiable type, e.g. 'follow' or 'joined'.

The notifiable ID, the thing the notification is about, e.g. a user or hackroom.

The read status, whether the notified user has seen this or not so we can set
it read/unread.