Display notifications
This creates the notifications controller and index view to display a users
notifications to them.

The link_through method marks notifications as read, which we'll use whenever
we load a notifications view.

At the moment we only have one type of notification so we've hard coded the
follow notification text for now.