Fix upcoming order event scope
Previously this was using the upcoming scope to fetch all future events,
thinking it ordered them but it doesn't. This meant the RSVP button on the
home page linked to the wrong event.

This creates a scope that takes upcoming events and puts them in order of
start date so the next event to start is first.