Rename events to meetups
I don't remember why I originally called these 'events' but it's a problem for
creating a news feed, where we'll create 'events' for users that can be
anything from joining a hackroom to adding a language to their profile.

This renames every instance of the word event in relation to meetups.

Obviously this is where more substantial specs would come in really useful for
spotting bugs that arise from this, and I should have done this over a series
of smaller commits but I don't have time right now.