Move welcome email from user model to registrations controller
Sending a welcome email when we create a user isn't the best idea in case we
ever need to recreate a user or create one manually but not send them a
welcome email. The email should be sent on the basis that they've registered
themselves.

There's also a method called in after actions that will be deprecated in
future versions of Rails, so this commit fixes that.