Add author to languages to generalise @commentable notification
When we create a comment on a post, we notify the creator of the post. We've
dodged this for languages because they don't have an author, but to make the
code more flexible and because whoever adds languages will want to know when
someone's commented on it (and in future anyone who subscribes to that
language) it makes sense to have a language author.

Also added a cheeky feature spec for administering languages to a) make sure
the author is saved and b) make sure the adding and editing of languages is
tested.