Fix database cleaner
'config.use_transactional_fixtures = true sets Rspec to clean the database, so
'despite setting up database_cleaner, it wasnt being used. Database cleaning
'broke in the last few commit and Im not sure why, maybe it was switching to
'Puma, maybe it was upgrading to Rails 5.

Either way, disabling this and allowing database_cleaner has fixed this.

I updated Rspec while debugging this and nothing broke so I'm keeping that.