Create custom comment feed for profile discussion
The discussion section wasn't loading previously because it was trying to load
the 'new comment' form, which it doesn't need and is set in the controller for
views where it is needed.

This fixes that by not using the 'shared/comments' partial, and passing in the
'follow_link: false' local so it renders a 'See comment' link to the comment
rather than the follower count of the comment author.

This also fixes the comment reply path_to_comment method, which was sending
you to the post path if the comment's commentable is a comment, but actually
should just send you to the parent comment.