Create Slack Invitations
We can use the Slack API to send members an invite to Slack rather than using
an app someone else made that doesn't authenticate users.

The risk of using that app is that people may spam invites and get our Slack
API blocked.

This also allows us to track who sends themselves invites, and encourage sign
up on the site.